+++ inst/scripts/f_predict.R
@@ -9,10 +9,31 @@ df = data.frame(t(x))
 colnames(df) <- bnames
 idx = which(complete.cases(df))
 df = df[idx,]
-args$newdata = df
+if ("model_fit" %in% class(model)) {
+  args$new_data = df # tidymodels fix
+} else {
+  args$newdata = df
+}
 args$object = model
 p = do.call("predict", args)
-out = array(NA, dim=c(1,d[2:4]))
-out[idx] = p
+out = array(NA, dim=c(length(output_names),nobs))
+if (is.list(p)) {
+  for (i in 1:length(output_names)) {
+    z = p[[output_names[i]]]
+    if (!is.numeric(z) && is.factor(z)) {
+      stop("Unexpected output from predict().")
+    }
+    out[i,idx] = as.numeric(z)
+  }
+} else if (is.matrix(p)) {
+  for (i in 1:length(output_names)) {
+    out[i,idx] = p[,output_names[i]]
+  }
+}  else if (is.numeric(p) || is.factor(p)) {
+  out[1, idx] = p
+} else {
+  stop("Unexpected output from predict().")
+}
+dim(out) <- c(length(output_names),d[2:4])
 write_chunk_from_array(out)
 #}