+++ src/gdalcubes/src/stream_apply_pixel.h
@@ -63,7 +63,12 @@ class stream_apply_pixel_cube : public cube {
             if (!_names.empty())
                 name = _names[i];
             else
-                name = "x" + std::to_string(i + 1);
+                name = "X" + std::to_string(i + 1);
+            
+            if (!std::isalnum(name[0])) {
+              GCBS_WARN("Variable name '" + name  + "' is not compatible with netCDF format; replacing with 'X" + name + "'");
+              name = "X" + name;
+            }
             _bands.add(band(name));
         }
     }