+++ src/gdalcubes/src/stream_reduce_space.h
@@ -65,6 +65,12 @@ class stream_reduce_space_cube : public cube {
                 name = _names[i];
             else
                 name = "band" + std::to_string(i + 1);
+            
+            if (!std::isalnum(name[0])) {
+              GCBS_WARN("Variable name '" + name  + "' is not compatible with netCDF format; replacing with 'X" + name + "'");
+              name = "X" + name;
+            }
+            
             _bands.add(band(name));
         }
     }